mptcp: Fix Issue #37: Explicitly cast srtt to 64-bit
We may have an integer overflow when multiplying the srtt's. C does not
do an automatic cast of 32-bit integers to 64-bit integers. So, we have
to make it explicit to avoid the integer overflow.

Fixes: Bug #37 (https://github.com/multipath-tcp/mptcp/issues/37)

+++ net/mptcp/mptcp_coupled.c
@@ -113,7 +113,7 @@ static void mptcp_ccc_recalc_alpha(struct sock *sk)
 		 * tmp will be in u64.
 		 */
 		tmp = div64_u64(mptcp_ccc_scale(sub_tp->snd_cwnd,
-				alpha_scale_num), sub_tp->srtt * sub_tp->srtt);
+				alpha_scale_num), (u64)sub_tp->srtt * sub_tp->srtt);
 
 		if (tmp >= max_numerator) {
 			max_numerator = tmp;
